Fix undefined behaviour in projectile and do_digging_projectile
Both out of bounds accesses due to missing isok checks.

+++ src/projectile.c
@@ -489,7 +489,8 @@ projectile(
 		}
 		/* space ahead has iron bars and no monster */
 		/* 1/5 chance for 'small' objects (see hits_bars), unless if launched from right beside the iron bars */
-		if (levl[bhitpos.x + dx][bhitpos.y + dy].typ == IRONBARS &&
+		if (isok(bhitpos.x + dx, bhitpos.y + dy) &&
+			levl[bhitpos.x + dx][bhitpos.y + dy].typ == IRONBARS &&
 			!m_at(bhitpos.x + dx, bhitpos.y + dy) && 
 			hits_bars(
 				/* object fired   */ thrownobj_p,
@@ -629,8 +630,13 @@ do_digging_projectile(
 	boolean youagr = (magr && (magr == &youmonst));
 	int newx = bhitpos.x + dx;
 	int newy = bhitpos.y + dy;
-	struct rm *room = &levl[newx][newy];
 	boolean shopdoor = FALSE, shopwall = FALSE;
+
+	/* Location must be valid */
+	if (!isok(newx, newy))
+		return;
+
+	struct rm *room = &levl[newx][newy];
 
 	/* Projectile must be a digger */
 	if (!((thrownobj->otyp == BLASTER_BOLT || thrownobj->otyp == HEAVY_BLASTER_BOLT || thrownobj->otyp == CARCOSAN_BOLT || thrownobj->otyp == LASER_BEAM)))
